cli/requirements.txt
-c constraints.txt
argh
pyserial
ptyprocess